Fix code which needed a 32 bit int to not use int/long
On amd64, but not on i386, tome produces LUA errors at various
times. The errors occur at savefile load, when wileding a weapon,
unweilding a weapon, and for other events. The same homedir, same
savefiles, same config, same etc., does not produce errors in the
32bit build of the package.

The error seems to be that the type Number needs to be a 32bit wide
integer; and seems to be defined by default as an long, with an option
to define it as an int for amd64. However, for the 11 arches Debian
supports, this would lead to a morass of selecting architectures in
the makefile to override the type; a better solution is to use
stdint.h and use a type which is 32 bits wide by design: int32_t.

This fixes the LUA errors, and Closes: Bug#419825

Signed-off-by: Manoj Srivastava <srivasta@debian.org>